// src/schema/bkk.d.ts
export interface paths {
// ...
	'/{dialect}/api/where/arrivals-and-departures-for-stop': {
		parameters: {
			query?: never;
			header?: never;
			path?: never;
			cookie?: never;
		};
		get: operations['getArrivalsAndDeparturesForStop'];
		put?: never;
		post?: never;
		delete?: never;
		options?: never;
		head?: never;
		patch?: never;
		trace?: never;
	};
// ...
}
export type webhooks = Record;
export interface components {
	schemas: {
// ...
		TransitTrip: {
			/**
			 * @description A menet azonosítója.
			 * @example BKK_C313504617
			 */
			id?: string;
			/**
			 * @description A menethez tartozó járat azonosítója.
			 * @example BKK_3040
			 */
			routeId?: string;
			/**
			 * @description A menet útvonalvezetésének azonosítója.
			 * @example BKK_C313504617
			 */
			shapeId?: string;
			/**
			 * @description A menethez tartozó GTFS block_id, amennyiben létezik.
			 * @example C31350_3040_5_15
			 */
			blockId?: string | null;
			/**
			 * @description A menet célállomása.
			 * @example Széll Kálmán tér M
			 */
			tripHeadsign?: string;
			/**
			 * @description A menet rövid neve.
			 * @example 4
			 */
			tripShortName?: string | null;
			/**
			 * @description A menet közlekedési naptárának azonosítója.
			 * @example BKK_C31350K2HCPER011
			 */
			serviceId?: string;
			/**
			 * @description A menet irányának azonosítója.
			 * @example 1
			 */
			directionId?: string;
			/**
			 * @description Szállítható-e kerékpár a meneten.
			 * @example false
			 */
			bikesAllowed?: boolean;
			/**
			 * @description Alacsonypadlós-e a menet.
			 * @example true
			 */
			wheelchairAccessible?: boolean;
		};
// ...
		TransitScheduleStopTime: {
			/**
			 * @description A megálló azonosítója
			 * @example BKK_050185
			 */
			stopId?: string;
			/**
			 * @description A megállóban kijelzett célállomás.
			 * @example Havanna utcai lakótelep
			 */
			stopHeadsign?: string;
			/**
			 * Format: int64
			 * @description A megállóba érkezés tervezett ideje epoch másodbercben. Az első megállónál hiányzik.
			 * @example 1625732820
			 */
			arrivalTime?: number | null;
			/**
			 * Format: int64
			 * @description A megállóból való indulás tervezett ideje epoch másodbercben. Az utolsó megállónál hiányzik.
			 * @example 1625732820
			 */
			departureTime?: number | null;
			/**
			 * Format: int64
			 * @description A megállóba érkezés becsült ideje epoch másodbercben, ha a járathoz van valós idejű adat. Az első megállónál hiányzik.
			 * @example 1625732820
			 */
			predictedArrivalTime?: number | null;
			/**
			 * Format: int64
			 * @description A megállóból való becsült indulás ideje epoch másodbercben, ha a járathoz van valós idejű adat. Az utolsó megállónál hiányzik.
			 * @example 1625732820
			 */
			predictedDepartureTime?: number | null;
			/**
			 * @description Igaz, ha a menethez tartozó valós idejű adatok bizonytalanok.
			 * @example true
			 */
			uncertain?: boolean | null;
			/**
			 * @description A bejegyzést tartalmazó menet azonosítója.
			 * @example BKK_C32269235
			 */
			tripId?: string;
			/**
			 * @description A bejegyzéshez tartozó menet menetrendi napja.
			 * @example 20210708
			 */
			serviceDate?: string;
			/**
			 * @description A bejegyzéshez tartozó menet alacsonypadlóssága.
			 * @example true
			 */
			wheelchairAccessible?: boolean;
			/**
			 * @description Igaz, ha a menet legalább egy rákövetkező megállója foglalást igényel.
			 * @example true
			 */
			mayRequireBooking?: boolean;
			/**
			 * @deprecated
			 * @description A megálló csoportjának azonosítója.
			 * @example [
			 *       "Havanna utcai lakótelep"
			 *     ]
			 */
			groupIds?: string[];
			/**
			 * @description A megállóhoz tartozó aktív zavarok azonosítóinak listája.
			 * @example [
			 *       "BKK_bkkinfo-75685"
			 *     ]
			 */
			alertIds?: string[];
		};
		ArrivalsAndDeparturesForStopOTPMethodResponse: {
			/**
			 * Format: int64
			 * @description Az aktuális szerveridő ezredmásodpercben.
			 * @example 1625747515786
			 */
			currentTime?: number;
			/**
			 * Format: int32
			 * @description A válasz API verziója.
			 * @example 3
			 */
			version?: number;
			status?: components['schemas']['Status'];
			/**
			 * Format: int32
			 * @description A válasz státusz kódja.
			 * @example 200
			 */
			code?: number;
			/**
			 * @description A válasz szövege.
			 * @example OK
			 */
			text?: string;
			data?: components['schemas']['TransitEntryWithReferencesTransitArrivalsAndDepartures'];
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}
export type $defs = Record;
export interface operations {
	searchAlerts: {
// ...
		responses: {
			/** @description Visszatér a `query` paraméterben megadott keresési feltételnek megfelelő zavarokkal, amik aktívak a megadott időintervallumnban. */
			200: {
				headers: {
					[name: string]: unknown;
				};
				content: {
					'application/json': components['schemas']['AlertSearchMethodResponse'];
				};
			};
		};
	};
// ...
}